Remove `SilvaClusterFinder` and `SilvaClusterFinderNtad` implementations and replace them with `BpcOldRunner` for building pure clusters.
