+++ tetrad-lib/src/main/java/edu/cmu/tetrad/search/Mimbuild.java
@@ -26,6 +26,7 @@ import edu.cmu.tetrad.data.ICovarianceMatrix;
 import edu.cmu.tetrad.data.Knowledge;
 import edu.cmu.tetrad.graph.*;
 import edu.cmu.tetrad.search.score.SemBicScore;
+import edu.cmu.tetrad.search.work_in_progress.BpcOld;
 import edu.cmu.tetrad.util.Matrix;
 import org.apache.commons.math3.distribution.ChiSquaredDistribution;
 import org.apache.commons.math3.optim.InitialGuess;
@@ -61,7 +62,7 @@ import java.util.List;
  *
  * @author josephramsey
  * @version $Id: $Id
- * @see Bpc
+ * @see BpcOld
  * @see Fofc
  * @see #getLatentsCov()
  * @see Boss
